Remove feature tags and former home page. Fix warnings
