+++ frontend/.env.sample
@@ -31,12 +31,8 @@ ALWAYS_DISPLAY_DASHBOARD_INFO=true
 ENABLE_DASHBOARDS=true
 DISABLE_PROFILES=false
 ENABLE_LEGACY_TOOL_SEARCH=false
-USE_CANVAS_MAP=false
 ENABLE_VERSIONING=false
 ENABLE_TOOL_PANEL=false
 CONSOLIDATE_INSIGHTS=true
 ENABLE_COUNTRY_PROFILES=true
 ENABLE_LOGISTIC_LAYERS_TAB=true
-ENABLE_TOP_NAV_REDESIGN=true
-ENABLE_FOOTER_REDESIGN=true
-ENABLE_NEW_HOME=true
