+++ frontend/.eslintrc.json
@@ -65,9 +65,6 @@
     "ENABLE_INTERSECTION_OBSERVER": true,
     "ENABLE_VERSIONING": true,
     "ENABLE_TOOL_PANEL": true,
-    "ENABLE_TOP_NAV_REDESIGN": true,
-    "ENABLE_FOOTER_REDESIGN": true,
-    "ENABLE_NEW_HOME": true,
     "CONSOLIDATE_INSIGHTS": true,
     "ENABLE_COUNTRY_PROFILES": true,
     "ENABLE_LOGISTIC_LAYERS_TAB": true,
@@ -88,7 +85,6 @@
     "TRANSIFEX_API_KEY": true,
     "PERF_TEST": true,
     "REDUX_LOGGER_ENABLED": true,
-    "USE_CANVAS_MAP": true,
     "USE_SERVICE_WORKER": true,
     "NAMED_MAPS_ENV": true,
     "GFW_WIDGETS_BASE_URL": true,
