+++ frontend/.storybook/webpack.config.js
@@ -38,7 +38,6 @@ module.exports = async ({ config }) =>
         PDF_DOWNLOAD_URL: JSON.stringify(process.env.PDF_DOWNLOAD_URL),
         PERF_TEST: process.env.PERF_TEST === 'true',
         REDUX_LOGGER_ENABLED: process.env.REDUX_LOGGER_ENABLED === 'true',
-        USE_CANVAS_MAP: process.env.USE_CANVAS_MAP === 'true',
         USE_SERVICE_WORKER: process.env.USE_SERVICE_WORKER === 'true',
         TRANSIFEX_API_KEY: JSON.stringify(process.env.TRANSIFEX_API_KEY),
         HOME_VIDEO_ID: JSON.stringify(process.env.HOME_VIDEO_ID),
