+++ frontend/docs/frontend/environments.md
@@ -67,7 +67,6 @@ REDUX_LOGGER_ENABLED=
 
 # feature flags
 USE_SERVICE_WORKER=
-USE_CANVAS_MAP=
 ALWAYS_DISPLAY_DASHBOARD_INFO=
 DATA_DOWNLOAD_ENABLED=
 DATA_FORM_ENABLED=
