--- frontend/scripts/react-components/home/author-footer/author-footer.scss
@@ -1,41 +1,0 @@
-@import 'styles/settings';
-
-.c-author-footer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  height: 40px;
-  width: 100%;
-
-  .author-details {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    justify-content: flex-end;
-    color: $charcoal-grey-faded;
-    font-family: $font-family-1;
-    font-size: 13px;
-    letter-spacing: $letter-spacing-x-narrow;
-    text-align: right;
-    text-transform: uppercase;
-
-    > span {
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      display: block;
-      overflow: hidden;
-    }
-  }
-
-  .author-avatar {
-    flex-shrink: 0;
-    margin: 0 0 0 10px;
-    width: 40px;
-    height: 40px;
-    border-radius: 100%;
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-color: darken($light-gray, 12%);
-  }
-}
