--- frontend/scripts/react-components/home/banner/banner-styles.scss
@@ -1,57 +1,0 @@
-@import 'styles/settings';
-
-.c-banner {
-  border: solid 2px rgba($yellow-gray, 0.5);
-  box-shadow: $box-shadow;
-  position: relative;
-  width: 420px;
-  height: 190px;
-
-  .background-white {
-    background-color: white;
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-  }
-
-  .background-image {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-size: 104%;
-    background-position: 50%;
-    will-change: opacity;
-
-    @media only screen and (max-width: '400px') {
-      background-size: 135vw;
-    }
-  }
-}
-
-.banner-content {
-  display: flex;
-  flex-flow: column;
-  justify-content: space-between;
-  height: 100%;
-  padding: 25px 75px 20px 20px;
-
-  .banner-title {
-    font-family: $font-family-2;
-    font-size: $font-size-big;
-    letter-spacing: -0.5px;
-    line-height: 1.3;
-    font-weight: $font-weight-bold;
-  }
-
-  .banner-link {
-    font-family: $font-family-1;
-    font-size: $font-size-regular;
-    text-transform: uppercase;
-    font-weight: $font-weight-bold;
-    margin-top: 40px;
-  }
-}
