--- frontend/scripts/react-components/home/home-video/home-video.scss
@@ -1,16 +1,0 @@
-@import 'styles/../node_modules/plyr/src/sass/plyr';
-
-.plyr {
-  visibility: hidden;
-  height: 1px;
-  width: 1px;
-  margin-left: -1px; // otherwise video is not playing after first click on the button (only iOS Safari ofc) dunno why???
-  min-width: 0;
-
-  &.-is-fullscreen {
-    visibility: visible;
-    height: 100%;
-    width: 100%;
-    margin-left: 0;
-  }
-}
