--- frontend/scripts/react-components/home/home.initial-state.js
@@ -1,17 +1,0 @@
-export default {
-  tweets: [
-    /**
-     * { id, text, screen_name, created_at, image_url }
-     */
-  ],
-  posts: [
-    /**
-     * { title, titleColor, description, date, imageUrl, highlighted, completePostUrl }
-     */
-  ],
-  testimonials: [
-    /**
-     * { quote, authorName, authorTitle, imageUrl }
-     */
-  ]
-};
