+++ frontend/scripts/react-components/home/home.register.js
@@ -1,7 +1,3 @@
-import reducerRegistry from 'reducer-registry';
-import reducer from './home.reducer';
 import * as homeActions from './home.actions';
-
-reducerRegistry.register('home', reducer);
 
 export { homeActions };
