--- frontend/scripts/react-components/home/home.thunks.js
@@ -1,6 +1,0 @@
-import { homeActions } from 'react-components/home/home.register';
-
-export const getPostsContent = dispatch => dispatch(homeActions.getHomeContent('posts'));
-export const getTestimonialsContent = dispatch =>
-  dispatch(homeActions.getHomeContent('testimonials'));
-export const getTweetsContent = dispatch => dispatch(homeActions.getHomeContent('tweets'));
