+++ frontend/scripts/react-components/home/homepage.scss
@@ -15,55 +15,4 @@
       background: $light-gray;
     }
   }
-
-  .homepage-map {
-    margin: 60px 0 0;
-
-    .homepage-map-sentence-selector {
-      margin: 0 0 25px;
-    }
-
-    .homepage-map-container {
-      margin: 0 -1.25rem;
-      width: 100vw;
-
-      @media only screen and (min-width: $breakpoint-foundation-small) {
-        margin: 0 auto;
-        width: 100%;
-        max-width: 700px;
-      }
-    }
-
-    .homepage-map-link-container {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      margin: 0 0 60px;
-    }
-
-    &.-bottom {
-      background-color: $white;
-      margin: 0;
-      padding: 60px 0 0;
-    }
-  }
-
-  .sliders {
-    background-color: $white;
-    padding: 60px 0;
-
-    &.-banner {
-      @media only screen and (max-width: $breakpoint-foundation-small) {
-        padding: 120px 0;
-      }
-    }
-
-    @media only screen and (max-width: $breakpoint-foundation-small) {
-      padding: 40px 0;
-    }
-
-    .c-slider-section:first-of-type {
-      padding: 0;
-    }
-  }
 }
