--- frontend/scripts/react-components/home/twitter-feed/twitter-feed.scss
@@ -1,102 +1,0 @@
-@import 'styles/settings';
-
-.-is-legacy-browser .c-twitter-feed {
-  left: -70px;
-}
-
-.c-twitter-feed {
-  position: relative;
-  width: 350px;
-
-  @media only screen and (max-width: $breakpoint-foundation-small) {
-    display: none;
-  }
-}
-
-.twitter-token {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 40px;
-  width: 40px;
-  background-color: $strong-pink;
-  border-radius: 100%;
-  margin: 0 0 0 12px;
-
-  .icon-twitter {
-    fill: $white;
-    margin-top: 2px;
-    margin-left: 1px;
-  }
-
-  @media only screen and (max-width: $breakpoint-foundation-small) {
-    display: none;
-  }
-}
-
-.tweet-box {
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  display: flex;
-  flex-shrink: 0;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 200px;
-  padding: 15px 20px 20px;
-  background-color: $white;
-  border-bottom: solid 3px $strong-pink;
-  box-shadow: 0 10px 40px 0 rgba(52, 68, 76, 0.25);
-  opacity: 0;
-  transform: translateX(40px);
-  transition: all 400ms $ease-in-out-cubic;
-
-  &::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    width: 9px;
-    height: 9px;
-    transform: translate(50%, calc(-50% - 9px)) rotate(-45deg);
-    background-color: $white;
-  }
-
-  &.-entering {
-    opacity: 0.5;
-    transform: translateX(0);
-    transition: all 400ms $ease-in-out-cubic 200ms;
-  }
-
-  &.-entered {
-    opacity: 1;
-    transform: translateX(0);
-  }
-
-  &.-exiting {
-    opacity: 0;
-    transform: translateX(0) scale(0.8, 0.8);
-  }
-
-  &.-exited {
-    transform: translateX(40px) scale(1, 1);
-    z-index: -1;
-  }
-}
-
-.-is-legacy-browser .tweet-box-content {
-  width: 25vw;
-}
-
-.tweet-box-content {
-  height: calc(100% - 40px + 20px);
-  color: $charcoal-grey;
-  font-family: $font-family-2;
-  font-size: 14px;
-  font-weight: 300;
-  line-height: 1.8;
-
-  > a {
-    color: $medium-pink;
-  }
-}
