--- frontend/scripts/react-components/new-home/entrypoints/entrypoints.scss
@@ -1,60 +1,0 @@
-@import 'styles/settings';
-$entrypoint-slide-height: 550px;
-$entrypoints-stacked-card-size: 360px;
-
-.c-new-entrypoints {
-  height: 100%;
-  padding-top: $nav-height;
-
-  @media screen and (min-width: $breakpoint-mobile) {
-    padding-top: $nav-height * 2;
-  }
-
-  .column {
-    padding: 0;
-  }
-
-  .row-link {
-    height: $entrypoints-stacked-card-size
-  }
-
-  .entrypoint-slide {
-    position: relative;
-    height: $entrypoints-stacked-card-size;
-    background-position: bottom right;
-    background-size: 75% auto;
-    background-repeat: no-repeat;
-    background-color: $light-gray;
-
-    &.-profile {
-      background-color: $egg-shell;
-    }
-
-    @media screen and (min-width: $breakpoint-mobile) {
-      background-size: 60% auto;
-    }
-  }
-
-  .entrypoint-slide-content {
-    position: relative;
-    padding: 45px 20px 0 20px;
-    height: 100%;
-    width: 100%;
-    z-index: $z-above;
-  }
-
-  .entrypoint-content {
-    position: relative;
-    padding: 45px 20px 0 20px;
-    display: flex;
-    height: 100%;
-    width: 100%;
-    z-index: $z-above;
-  }
-
-  .entrypoint-text {
-    max-width: 390px;
-    min-width: 30%;
-    margin-right: 50px;
-  }
-}
