--- frontend/scripts/react-components/new-home/home.actions.js
@@ -1,6 +1,0 @@
-export const HOME__CLICK_ENTRYPOINT = 'HOME__CLICK_ENTRYPOINT';
-
-export const clickEntrypoint = link => ({
-  type: HOME__CLICK_ENTRYPOINT,
-  payload: link
-});
