--- frontend/scripts/react-components/new-home/home.register.js
@@ -1,3 +1,0 @@
-import * as homeActions from './home.actions';
-
-export { homeActions };
