--- frontend/scripts/react-components/new-home/homepage.scss
@@ -1,18 +1,0 @@
-@import 'styles/settings';
-
-.c-new-homepage {
-  overflow: hidden;
-
-  .homepage-entrypoints {
-    position: relative;
-    background: linear-gradient(to bottom, $light-gray 0%, $light-gray 33.33%, $egg-shell 33.33%, $egg-shell 66.66%, $light-gray 66.66%, $light-gray 100%);
-
-    @media only screen and (min-width: $home-entrypoint-breakpoint) {
-      background: $light-gray
-    }
-
-    &.-hide-profiles {
-      background: $light-gray;
-    }
-  }
-}
