+++ frontend/scripts/react-components/shared/hero/hero.scss
@@ -9,10 +9,6 @@
     height: 400px;
     overflow: hidden;
 
-    .hero-content {
-      display: none;
-    }
-
     @media screen and (max-width: $breakpoint-foundation-small) {
       height: 190px;
     }
@@ -27,197 +23,12 @@
     top: 0;
   }
 
-  .hero-logo-container {
-    height: 38px;
-    width: 210px;
-    background-size: contain;
-    background-repeat: no-repeat;
-
-    @media screen and (max-width: $breakpoint-foundation-small) {
-      height: 34px;
-    }
-
-    .hero-trase-logo {
-      height: 100%;
-      width: 100%;
-      background-size: contain;
-      background-repeat: no-repeat;
-    }
-  }
-
-  .hero-title {
-    font-family: $font-family-2;
-    font-size: 46px;
-    letter-spacing: -1px;
-    line-height: 1.3;
-    font-weight: 300;
-    color: $white;
-    width: 590px;
-    max-width: 100%;
-    margin: 32px 0 25px;
-
-    @media only screen and (max-width: $breakpoint-foundation-small) {
-      font-size: $font-size-big;
-      line-height: 1.67;
-      margin: 5px 0 40px;
-    }
-  }
-
-  .hero-play-container {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    font-family: $font-family-1;
-    color: $white;
-    font-size: 13px;
-    letter-spacing: -.8px;
-
-    .hero-play-button {
-      position: relative;
-      width: 40px;
-      height: 40px;
-      border-radius: 100%;
-      background-color: $medium-pink;
-      cursor: pointer;
-
-      &::after {
-        content: '';
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-30%, -50%);
-        background: transparent;
-        width: 0;
-        height: 0;
-        border-color: transparent transparent transparent $white;
-        border-style: solid;
-        border-width: 5px 0 5px 8px;
-      }
-
-      &:hover {
-        background-color: darken($medium-pink, 4%);
-      }
-    }
-
-    > span {
-      text-transform: uppercase;
-      margin-left: 10px;
-      width: 120px;
-    }
-  }
-
   > .row {
     height: 100%;
     position: relative;
 
     @media screen and (max-width: $breakpoint-foundation-small) {
       padding-top: 50px;
-    }
-  }
-
-  .layover {
-    display: flex;
-    justify-content: flex-end;
-    align-items: flex-end;
-    position: absolute;
-    bottom: -13px;
-    right: 0;
-    max-width: 50%;
-    z-index: $z-above-hero;
-    display: flex;
-
-    &.-banner {
-      margin-right: 20px;
-
-      @media only screen and (max-width: $breakpoint-foundation-small) {
-        position: relative;
-        right: initial;
-        width: 100%;
-        z-index: 3;
-        max-width: 100%;
-      }
-    }
-
-    &.-tweets {
-      width: 350px;
-    }
-  }
-
-  .story-box {
-    position: relative;
-    display: flex;
-    width: 350px;
-    height: 320px;
-    flex-direction: column;
-    background-color: $white;
-    border-bottom: solid 3px $strong-pink;
-    box-shadow: 0 10px 40px 0 rgba(52, 68, 76, 0.25);
-    overflow: hidden;
-
-    @media screen and (max-width: $breakpoint-foundation-small) {
-      display: none;
-    }
-
-    .story-box-image {
-      width: 100%;
-      height: 180px;
-      background-color: darken($light-gray, 12%);
-      background-size: cover;
-    }
-
-    .story-box-close {
-      position: absolute;
-      top: 5px;
-      right: 5px;
-      height: 18px;
-      width: 18px;
-      cursor: pointer;
-
-      &:before, &:after {
-        content: '';
-        position: absolute;
-        top: 0;
-        height: 16px;
-        width: 2px;
-        background-color: $charcoal-grey;
-        border: 0;
-        cursor: pointer;
-
-        &:hover {
-          background-color: lighten($charcoal-grey, 12%);
-        }
-      }
-
-      &:before {
-        transform: rotate(45deg);
-      }
-      &:after {
-        transform: rotate(-45deg);
-      }
-    }
-
-    .story-box-content {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      align-items: flex-end;
-      height: 140px;
-      padding: 20px;
-    }
-
-    .story-box-title {
-      width: 100%;
-      color: $gray-text;
-      font-family: $font-family-2;
-      font-size: 20px;
-      font-weight: 300;
-      line-height: 1.6;
-      text-transform: capitalize;
-    }
-
-    .story-box-link {
-      text-align: right;
-      cursor: pointer;
     }
   }
 }
