--- frontend/scripts/react-components/shared/quote-card/quote-card.scss
@@ -1,38 +1,0 @@
-@import 'styles/settings';
-
-$slide-height-small: 300px;
-
-.c-quote-card {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-  height: $slide-height-small;
-
-  .quote-card-container {
-    height: 100%;
-    width: 100%;
-    padding: 20px;
-    background-color: $light-gray;
-
-    .quote-card-content {
-      height: calc(100% - 40px);
-      overflow: hidden;
-      font-family: $font-family-2;
-      font-size: 16px;
-      font-weight: 300;
-      font-style: italic;
-      line-height: 1.6;
-      letter-spacing: -0.5px;
-      color: $gray-text;
-
-      > a {
-        font-weight: $font-weight-regular;
-      }
-    }
-  }
-
-  &:first-child {
-    margin: 0;
-  }
-}
